refactor(quality): split adobeFireflyClient.ts god-file into nine leaf modules
2,952 lines -> 693, with nine new modules that each land under the 1000-line cap
in this same PR (check-file-size.mjs rejects a new file above the cap outright,
so a staged oversized intermediate was never an option):

  adobeFireflyCatalog.ts      282  endpoint URLs, tunables, image/video model catalog
  adobeFireflyCredentials.ts  205  IMS token / cookie-blob recognition and extraction
  adobeFireflyPayload.ts      479  model resolution and generate-request payloads
  adobeFireflyArp.ts          382  nonces, ARP session ids, request headers
  adobeFireflyUpload.ts       347  source-image extraction and storage upload
  adobeFireflyResponses.ts    230  upstream response parsing, job-status helpers
  adobeFireflyIms.ts          256  IMS token exchange and access-token resolution
  adobeFireflyAccount.ts      157  credits balance and model discovery
  adobeFireflyPoll.ts         155  async job polling
  adobeFireflyClient.ts       693  generateImage/generateVideo + barrel

Follows the decomposition already established for this family: Security,
ModelSnapshot, References, Upscale, Models, Session and BrowserLogin are all
siblings of the same client.

Behavior-preserving by construction. The barrel re-exports every symbol
EXPLICITLY BY NAME rather than with `export *`, so the public surface is
unchanged even though nine previously module-private helpers had to become
`export` to cross the new boundaries -- those nine stay out of the barrel and
remain invisible to the 15 files that import this module, none of which needed
an edit.

Dependency graph between the leaves is acyclic:
  catalog, credentials -> (none)
  payload -> catalog
  arp, ims -> catalog, credentials
  upload, responses -> catalog, credentials, arp
  account, poll -> catalog, credentials, responses
  client -> all of the above
